requirements.txt
Flask==3.0.3
google-api-python-client==2.125.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.5.4
gspread==6.1.2
gunicorn==22.0.0
joblib==1.3.2
linkedin-api==2.1.1
linkedin_api==2.1.1
matplotlib==3.9.0
pandas==2.2.2
python-docx==1.1.2
requests==2.32.2
scikit-learn==1.5.0
